Trying to add title kwarg to show
